[#29] add service package and move service interface to
