+++ core/src/main/java/com/ztkmkoo/dss/core/actor/rest/DssRestMasterActor.java
@@ -4,6 +4,7 @@ import akka.actor.typed.ActorRef;
 import akka.actor.typed.Behavior;
 import akka.actor.typed.javadsl.ActorContext;
 import akka.actor.typed.javadsl.Behaviors;
+import com.ztkmkoo.dss.core.actor.rest.service.DssRestActorService;
 import com.ztkmkoo.dss.core.message.rest.*;
 import io.netty.handler.codec.http.HttpResponseStatus;
