+++ core/src/main/java/com/ztkmkoo/dss/core/network/rest/handler/DssRestChannelInitializer.java
@@ -5,8 +5,8 @@ import akka.actor.typed.Behavior;
 import akka.actor.typed.javadsl.ActorContext;
 import akka.actor.typed.javadsl.Behaviors;
 import com.ztkmkoo.dss.core.actor.exception.DssUserActorDuplicateBehaviorCreateException;
-import com.ztkmkoo.dss.core.actor.rest.DssRestActorService;
 import com.ztkmkoo.dss.core.actor.rest.DssRestMasterActor;
+import com.ztkmkoo.dss.core.actor.rest.service.DssRestActorService;
 import com.ztkmkoo.dss.core.message.rest.DssRestChannelInitializerCommand;
 import com.ztkmkoo.dss.core.message.rest.DssRestChannelInitializerCommandHandlerUnregistered;
 import com.ztkmkoo.dss.core.message.rest.DssRestMasterActorCommand;
