+++ core/src/test/java/com/ztkmkoo/dss/core/actor/rest/DssRestMasterActorTest.java
@@ -5,6 +5,7 @@ import akka.actor.typed.ActorRef;
 import com.ztkmkoo.dss.core.actor.AbstractDssActorTest;
 import com.ztkmkoo.dss.core.actor.rest.entity.DssRestServiceRequest;
 import com.ztkmkoo.dss.core.actor.rest.entity.DssRestServiceResponse;
+import com.ztkmkoo.dss.core.actor.rest.service.DssRestActorService;
 import com.ztkmkoo.dss.core.message.rest.DssRestChannelHandlerCommand;
 import com.ztkmkoo.dss.core.message.rest.DssRestChannelHandlerCommandResponse;
 import com.ztkmkoo.dss.core.message.rest.DssRestMasterActorCommand;
